Fix tests. Fix label association for Status field, improve CustomSelect accessibility, and update loading text matcher in tests
